Implement random for `FieldElement` & `Scalar` generrically.
Generate a valid result choosen uniformly using user-
provided rng.
By `rng` we mean any Rng that implements: `Rng` + `CryptoRng`.
